mt4tradingbot/requirements.txt
alembic==1.6.4
amqp==5.0.6
aniso8601==9.0.1
asgiref==3.3.4
attrs==21.2.0
billiard==3.6.4.0
celery==5.1.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
colorama==0.4.4
cors==1.0.1
cryptography==3.4.7
filelock==3.0.12
Flask==2.0.0
Flask-Cors==3.0.10
Flask-Migrate==2.5.2
flask-restx==0.4.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
future==0.18.2
gevent==21.1.2
greenlet==1.1.0
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
jsonschema==3.2.0
kombu==5.1.0
Mako==1.1.4
MarkupSafe==2.0.1
prompt-toolkit==3.0.18
psycopg2-binary==2.8.6
pyaes==1.6.1
pyasn1==0.4.8
pycparser==2.20
PyJWT==2.1.0
pyrsistent==0.17.3
PySocks==1.7.1
python-dateutil==2.8.1
python-dotenv==0.17.1
python-editor==1.0.4
pytz==2021.1
redis==3.5.3
requests==2.25.1
requests-file==1.5.1
rsa==4.7.2
six==1.16.0
SQLAlchemy==1.4.15
Telethon==1.21.1
tldextract==3.1.0
urllib3==1.26.4
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.0.1
zope.event==4.5.0
zope.interface==5.4.0
